requirements.txt
beautifulsoup4
requests
pillow
python-dotenv
google-generativeai
